tests/wpt/tests/css/css-scoping/slotted-invalidation.html

<!doctype html>
<title>CSS Test: Style invalidation for ::slotted()</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link rel="author" title="Rune Lillesveen" href="mailto:futhark@chromium.org">
<link rel="author" title="Emilio Cobos Ãlvarez" href="mailto:emilio@crisal.io">
<link rel="help" href="https://drafts.csswg.org/css-scoping/#slotted-pseudo">
<div id="host">
  <div>
    <span></span>
    <span></span>
  </div>
  <div id="slotted">
    <span></span>
    <span></span>
  </div>
  <div>
    <span></span>
    <span></span>
  </div>
</div>
<script>
test(function() {
  var root = host.attachShadow({"mode":"open"});
  root.innerHTML = '<style>.outer ::slotted(#slotted) { background-color: red } .outer .inner::slotted(#slotted) { background-color: green }</style><div id="outer"><slot id="inner"></slot></div>';

  assert_equals(window.getComputedStyle(slotted).backgroundColor, "rgba(0, 0, 0, 0)");

  root.querySelector("#outer").className = "outer";
  assert_equals(window.getComputedStyle(slotted).backgroundColor, "rgb(255, 0, 0)");

  root.querySelector("#inner").className = "inner";
  assert_equals(window.getComputedStyle(slotted).backgroundColor, "rgb(0, 128, 0)");
})
</script>
