third_party/WebKit/LayoutTests/external/wpt/css/css-scoping/slotted-invalidation-expected.txt

This is a testharness.js-based test.
FAIL CSS Test: Style invalidation for ::slotted() assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
Harness: the test ran to completion.
